implement `getLintRulesFromConfig` tests
