+++ source-code/core/src/lint/linter.ts
@@ -1,10 +1,12 @@
 import type { Resource, Message, Pattern } from '../ast/schema.js'
 import type { Config, EnvironmentFunctions } from '../config/schema.js'
 import { createContext, LintedResource } from './context.js';
-import { getLintRulesFromConfig, LintableNode, LintableNodeByType, ConfiguredLintRule, NodeVisitor, NodeVisitors, TargetReferenceParameterTuple } from './rule.js';
+import type { LintableNode, LintableNodeByType, ConfiguredLintRule, NodeVisitor, NodeVisitors, TargetReferenceParameterTuple } from './rule.js';
 
 const getResourceForLanguage = (resources: Resource[], language: string) =>
 	resources.find(({ languageTag }) => languageTag.name === language);
+
+export const getLintRulesFromConfig = (config: Config) => (config.lint?.rules || []).flat()
 
 export const lint = async (config: Config, env: EnvironmentFunctions) => {
 	const { referenceLanguage, languages, readResources } = config
